requirements.txt
click
open_clip_torch
Pillow
PyYAML
safetensors
torch
tqdm
bayesian-optimization
streamlit
stqdm
dacite
